requirements.txt
pandas
numpy
scikit-learn
boto3
pymongo
python-dotenv
joblib

